Clarify raw data endpoint requirement
We have noticed that frequently candidates are confused as to whether the first requirement to consume raw data refers to the client doing the consumption or the server. While it could be argued that the bonus objective for bulk ingestion should disambiguate this, I feel it is more helpful to have uniform language with the second requirement ("expose").